Implemented multithreaded key-gen (not yet *fully* functional on MacOS)
The issue on MacOS is with my `mac_getrandom` implementation, which seems to not close every file descriptor it opens. Because of this, it tends to exceed the `ulimit` maximum open file descriptors number.

If you MUST run this on MacOS (while I work out the weird concurrency bug with `mac_getrandom`), just set `ulimit -n` to a much higher number than the MacOS default.